+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Color palettes' gradient color array;
 - The brands' official logos and icons downloaded from their official web sites.
 
+## [0.1.3] - 2021-04-13
+### Added [#19](https://github.com/rafamizes/flutter_brand_palettes/issues/19)
+- Material design's light green color palette;
+- Material design's light green accent color palette;
+- Material design's lime color palette;
+- Material design's lime accent color palette;
+- Material design's teal color palette;
+- Material design's teal accent color palette.
+
 ## [0.1.2] - 2021-04-12
 ### Added [#18](https://github.com/rafamizes/flutter_brand_palettes/issues/18)
 - Material design's amber color palette;